removed a print statement
